deleted old bottom sheet classes
