+++ android/anc/src/main/java/org/smartregister/fhircore/anc/ui/family/removefamilymember/RemoveFamilyMemberQuestionnaireActivity.kt
@@ -30,8 +30,8 @@ import org.smartregister.fhircore.anc.data.family.model.FamilyMemberItem
 import org.smartregister.fhircore.anc.ui.family.details.FamilyDetailsActivity
 import org.smartregister.fhircore.anc.util.bottomsheet.BottomSheetDataModel
 import org.smartregister.fhircore.anc.util.bottomsheet.BottomSheetHolder
-import org.smartregister.fhircore.anc.util.bottomsheet.BottomSheetListDialog
 import org.smartregister.fhircore.anc.util.bottomsheet.NewBottomSheetListDialog
+import org.smartregister.fhircore.anc.util.bottomsheet.OnClickedListItems
 import org.smartregister.fhircore.anc.util.othersEligibleForHead
 import org.smartregister.fhircore.engine.ui.questionnaire.QuestionnaireActivity
 import org.smartregister.fhircore.engine.util.extension.showToast
@@ -39,7 +39,7 @@ import org.smartregister.fhircore.engine.util.extension.toAgeDisplay
 
 @AndroidEntryPoint
 class RemoveFamilyMemberQuestionnaireActivity :
-  QuestionnaireActivity(), BottomSheetListDialog.OnClickedListItems {
+  QuestionnaireActivity(), OnClickedListItems {
 
   @VisibleForTesting lateinit var saveBtn: Button
   private lateinit var familyId: String
@@ -100,8 +100,6 @@ class RemoveFamilyMemberQuestionnaireActivity :
     val title = getString(R.string.label_assign_new_family_head)
     val listTitle = getString(R.string.label_select_new_head)
     val warning = getString(R.string.label_remove_family_warning)
-    // BottomSheetListDialog(this, BottomSheetHolder(title, listTitle, warning, options),
-    // this).show()
     NewBottomSheetListDialog(BottomSheetHolder(title, listTitle, warning, options), this)
       .show(supportFragmentManager, "")
   }
