+++ android/anc/src/main/res/values/strings.xml
@@ -140,7 +140,7 @@
     <string name="remove_family_member_warning">Removing this will remove their entire health record from your device. This action can’t be undone.</string>
     <string name="family_resource_not_found">"Family resource for that ID NOT Found"</string>
     <string name="label_remove_family_warning">Please assign a new family head, family nolonger has a family head.</string>
-    <string name="label_select_new_head">Select a new head</string>
+    <string name="label_select_new_head">Select a new family head</string>
     <string name="label_assign_new_family_head">Assign new family head</string>
     <string name="alert_message_abort_operation">Are you sure you want to abort this operation?</string>
     <string name="alert_title_abort">Abort</string>
